feat: textarea length 구현 - #28
