+++ src/components/writePage/ApplyAnswer.js
@@ -1,10 +1,14 @@
 import styled from 'styled-components';
 
-const ApplyAnswer = ({ register, name }) => {
+const ApplyAnswer = ({ register, name, value }) => {
   return (
-    // htmlFor, id는 따로 데이터 배열 고유값 id 만들어서 사용
-    <AnswerContainer htmlFor="answer">
-      <AnswerTextarea id="answer" {...register[name]}></AnswerTextarea>
+    <AnswerContainer htmlFor={name}>
+      <AnswerTextarea
+        id={name}
+        {...register(name)}
+        autoFocus={false}
+      ></AnswerTextarea>
+      <AnswerLength>( {value[name].length} / 1000 )</AnswerLength>
     </AnswerContainer>
   );
 };
@@ -12,21 +16,38 @@ const ApplyAnswer = ({ register, name }) => {
 const AnswerContainer = styled.label`
   width: 330px;
   height: 330px;
-  padding: 26px 27px;
+  padding: 26px 27px 40px 27px;
   margin-bottom: 50px;
   border-radius: 15px;
   background-color: ${({ theme }) => theme.colors.CARD_BG};
   @media ${({ theme }) => theme.devices.TABLET} {
     width: 560px;
     height: 1067px;
-    padding: 34px 35px;
+    padding: 36px 35px 56px 35px;
     margin-bottom: 110px;
   }
 
   @media ${({ theme }) => theme.devices.DESKTOP} {
     width: 972px;
     height: 882px;
-    padding: 62px 61px;
+    padding: 60px 61px 80px 61px;
+  }
+`;
+
+const AnswerLength = styled.p`
+  float: right;
+  color: #939393;
+  font-size: 10px;
+  font-weight: 500;
+  margin-top: 12px;
+  @media ${({ theme }) => theme.devices.TABLET} {
+    font-size: 14px;
+    margin-top: 16px;
+  }
+
+  @media ${({ theme }) => theme.devices.DESKTOP} {
+    ${({ theme }) => theme.typographies.DEFAULT_TXT}
+    margin-top: 26px;
   }
 `;
 const AnswerTextarea = styled.textarea`
