+++ src/components/writePage/UserInfoInput.js
@@ -12,9 +12,14 @@ const UserInfoInput = ({
     <Container>
       <InfoLabel>{label}</InfoLabel>
       {isDisabled ? (
-        <InfoInput disabled />
+        <InfoInput disabled error={errors[name]} />
       ) : (
-        <InfoInput placeholder={placeholder} id={name} {...register(name)} />
+        <InfoInput
+          placeholder={placeholder}
+          id={name}
+          {...register(name)}
+          error={errors[name]}
+        />
       )}
     </Container>
   );
@@ -53,11 +58,16 @@ const InfoInput = styled.input`
   height: 100%;
   padding: 8px 10px;
   border: none;
+  border: ${({ error, theme }) =>
+    error ? `1px solid ${theme.colors.HOVER_BTN}` : 'none'};
   border-radius: 8px;
   background-color: ${({ theme }) => theme.colors.CARD_BG};
   &:disabled {
     background-color: #202020;
     border: 1px solid ${({ theme }) => theme.colors.MODAL_BG};
+  }
+  &:focus {
+    border: ${({ theme }) => `2px solid ${theme.colors.MODAL_BG}`};
   }
   font-size: 10px;
   font-weight: 500;
